+++ pennylane/argmap.py
@@ -20,16 +20,15 @@ from pennylane import numpy as np
 
 class ArgMapError(Exception):
     r"""Exception raised by an :class:`~.pennylane.argmap.ArgMap` instance
-    when it is unable to access or write a requested item."""
+    when it is unable to create an instance, or access or write a requested item."""
 
 
 @lru_cache(maxsize=None)
 def _interpret_key(key, single_arg=False):
+    r"""Interpret ArgMap key into argument and parameter index."""
     if isinstance(key, tuple):
         if len(key) == 2:
-            if isinstance(key[1], tuple):
-                return key
-            if key[1] is None:
+            if isinstance(key[1], tuple) or key[1] is None:
                 return key
         if len(key) == 0:
             return None, None
@@ -44,13 +43,75 @@ def _interpret_key(key, single_arg=False):
 
 
 class ArgMap(dict):
+    r"""Argument index map for storing objects per QNode argument and parameter.
+
+    Args:
+        data (list[tuple] or dict or object): Data to store in the ArgMap.
+            If not a dict and ``single_object=False``, ``data`` must be possible 
+            to be parsed via ``dict(data)``.
+        single_arg (bool): Whether the ``ArgMap`` stores objects for a single argument.
+        single_object (bool): Whether the ``ArgMap`` stores only a single object.
+
+    An ``ArgMap`` acts as a dictionary with a restricted type of keys in the context of
+    a given ``QNode``. Each key takes the form ``tuple[int, tuple[int] or int]``.
+    The first ``int`` in the tuple describes the argument index of the ``QNode``,
+    while the second entry describes the index within that argument.
+    That is ``(2, (0, 3))`` describes the parameter in the ``0``th row and ``3``rd
+    column of the ``2``nd argument of a QNode. The second entry of the tuple may be an
+    ``int`` instead, indexing a one-dimensional array.
+
+    If the argument index (first entry) of a key is ``None``, this indicates that there is
+    only one argument in the QNode, and all keys of the ``ArgMap`` should have the form
+    ``(None, tuple[int] or int)``.
+    If the parameter index (second entry) of a key is ``None``, the respective argument
+    is a scalar and no other key with the same argument index can exist in the ``ArgMap``.
+    If both entries are ``None``, i.e. ``key=(None, None)``, the QNode only takes a single
+    scalar argument.
+
+    The main feature of an ``ArgMap`` is its flexible access: Users can specify only an
+    argument index or only a parameter index if this fully uniquely describes the
+    parameter.
+
+    **Example**
+    Suppose we have the following ``QNode`` instance:
+
+    .. code-block:: python
+        
+        dev = qml.device('default.qubit', wires=2)
+        @qml.qnode(dev)
+        def circuit(arg1, arg2, arg3):
+            qml.Rot(arg1, wires=0)
+            qml.RX(arg2, wires=0)
+            qml.Rot(arg3, wires=1)
+            return qml.expval(qml.PauliY(0) @ qml.PauliY(1))
+
+    Then we might want to set up an ``ArgMap`` to store the generators of the rotations in
+    this circuit, depending on the parameter that controls them:
+
+    .. code-block:: python
+
+        generators = qml.ArgMap({
+            (0, 0): qml.PauliZ(0), 
+            (0, 1): qml.PauliY(0), 
+            (0, 2): qml.PauliZ(0), 
+            1: qml.PauliX(0), 
+            (1, 0): qml.PauliZ(1), 
+            (1, 1): qml.PauliY(1), 
+            (1, 2): qml.PauliZ(1), 
+        })
+
+    Then the generators can be accessed by using the tuples ``(arg_index, par_index)`` as key
+    with ``par_index=None`` for the ``PauliX(0)`` generator. The latter may alternatively
+    be accessed directly via ``generators[1]``.
+    """
     def __init__(self, data, single_arg=False, single_object=False):
         self.single_arg = single_arg
-        _data = self._preprocess_data(data, single_object)
+        self.single_object = single_object
+        _data = self._preprocess_data(data)
         super().__init__(_data)
 
-    def _preprocess_data(self, data, single_object):
-        if single_object:
+    def _preprocess_data(self, data):
+        if self.single_object:
             return {(None, None): data}
 
         if not isinstance(data, dict):